+++ src/lib/tuya.ts
@@ -14,6 +14,7 @@ import type {
     KeyValue,
     KeyValueAny,
     KeyValueNumberString,
+    KeyValueString,
     ModernExtend,
     Publish,
     Range,
@@ -59,7 +60,88 @@ export const dataTypes = {
     bitmap: 5, // [ 1,2,4 bytes ] as bits
 };
 
-export const M8ProTuyaWeatherCondition = {
+export interface TuyaWeatherCondition {
+    sunny?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    heavy_rain?: number;
+    cloudy?: number;
+    sandstorm?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    light_snow?: number;
+    snow?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    freezing_fog?: number;
+    rainstorm?: number;
+    shower?: number;
+    dust?: number;
+    spit?: number;
+    sleet?: number;
+    yin?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    freezing_rain?: number;
+    rain?: number;
+    fog?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    heavy_shower?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    heavy_snow?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    heavy_downpour?: number;
+    blizzard?: number;
+    hailstone?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    snow_shower?: number;
+    haze?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    thunder_shower?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    floating_dust?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    thunder_and_lighting?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    dust_storm?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    light_shower?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    rain_and_snow?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    dust_bowl?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    ice_pellets?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    strong_dust_storms?: number;
+    sandy?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    light_to_moderate_rain?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    mostly_sunny?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    very_heavy_rain?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    ice_pod?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    light_to_moderate_snow?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    few_clouds?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    light_snow_showers?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    moderate_snow?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    icy_needles?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    thunderstorm_with_ice_pods?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    light_rain?: number;
+    thunder?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    moderate_rain?: number;
+    thunderstorm?: number;
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    moderate_to_heavy_rain?: number;
+}
+
+export const M8ProTuyaWeatherCondition: TuyaWeatherCondition = {
     sunny: 100,
     heavy_rain: 101,
     cloudy: 102,
@@ -84,6 +166,47 @@ export const M8ProTuyaWeatherCondition = {
     snow_shower: 130,
     haze: 140,
     thunder_shower: 143,
+};
+
+export const F3ProTuyaWeatherCondition: TuyaWeatherCondition = {
+    heavy_rain: 101,
+    thunderstorm: 102,
+    dust_storm: 103,
+    light_snow: 104,
+    snow: 105,
+    freezing_fog: 106,
+    shower: 108,
+    floating_dust: 109,
+    thunder_and_lighting: 110,
+    light_shower: 111,
+    rain: 112,
+    rain_and_snow: 113,
+    dust_bowl: 114,
+    ice_pellets: 115,
+    strong_dust_storms: 116,
+    sandy: 117,
+    light_to_moderate_rain: 118,
+    mostly_sunny: 119,
+    sunny: 120,
+    haze: 121,
+    heavy_shower: 123,
+    heavy_snow: 124,
+    very_heavy_rain: 125,
+    blizzard: 126,
+    ice_pod: 127,
+    light_to_moderate_snow: 128,
+    few_clouds: 129,
+    light_snow_showers: 130,
+    moderate_snow: 131,
+    cloudy: 132,
+    icy_needles: 133,
+    thunderstorm_with_ice_pods: 136,
+    freezing_rain: 137,
+    snow_shower: 138,
+    light_rain: 139,
+    thunder: 140,
+    moderate_rain: 141,
+    moderate_to_heavy_rain: 144,
 };
 
 export enum TuyaWeatherID {
@@ -339,33 +462,48 @@ const tuyaExposes = {
             .numeric(`power_factor_${phase}`, ea.STATE)
             .withUnit("%")
             .withDescription(`Instantaneous measured power factor (phase ${phase.toUpperCase()})`),
-    switchType: () => e.enum("switch_type", ea.ALL, ["toggle", "state", "momentary"]).withDescription("Type of the switch"),
+    switchType: () => e.enum("switch_type", ea.ALL, ["toggle", "state", "momentary"]).withDescription("Type of the switch").withCategory("config"),
     switchTypeCurtain: () =>
         e
             .enum("switch_type_curtain", ea.ALL, ["flip-switch", "sync-switch", "button-switch", "button2-switch"])
-            .withDescription("External switch type"),
-    backlightModeLowMediumHigh: () => e.enum("backlight_mode", ea.ALL, ["low", "medium", "high"]).withDescription("Intensity of the backlight"),
-    backlightModeOffNormalInverted: () => e.enum("backlight_mode", ea.ALL, ["off", "normal", "inverted"]).withDescription("Mode of the backlight"),
+            .withDescription("External switch type")
+            .withCategory("config"),
+    switchTypeButton: () =>
+        e.enum("switch_type_button", ea.ALL, ["release", "press"]).withDescription("Determines when the button actuates").withCategory("config"),
+    backlightModeLowMediumHigh: () =>
+        e.enum("backlight_mode", ea.ALL, ["low", "medium", "high"]).withDescription("Intensity of the backlight").withCategory("config"),
+    backlightModeOffNormalInverted: () =>
+        e.enum("backlight_mode", ea.ALL, ["off", "normal", "inverted"]).withDescription("Mode of the backlight").withCategory("config"),
     backlightModeOffOn: () => e.binary("backlight_mode", ea.ALL, "ON", "OFF").withDescription("Mode of the backlight").withCategory("config"),
     indicatorMode: () =>
         e.enum("indicator_mode", ea.ALL, ["off", "off/on", "on/off", "on"]).withDescription("LED indicator mode").withCategory("config"),
     indicatorModeNoneRelayPos: () =>
         e.enum("indicator_mode", ea.ALL, ["none", "relay", "pos"]).withDescription("Mode of the indicator light").withCategory("config"),
-    powerOutageMemory: () => e.enum("power_outage_memory", ea.ALL, ["on", "off", "restore"]).withDescription("Recover state after power outage"),
+    powerOutageMemory: () =>
+        e.enum("power_outage_memory", ea.ALL, ["on", "off", "restore"]).withDescription("Recover state after power outage").withCategory("config"),
     batteryState: () => e.enum("battery_state", ea.STATE, ["low", "medium", "high"]).withDescription("State of the battery"),
     doNotDisturb: () =>
         e
             .binary("do_not_disturb", ea.STATE_SET, true, false)
-            .withDescription("Do not disturb mode, when enabled this function will keep the light OFF after a power outage"),
+            .withDescription("Do not disturb mode, when enabled this function will keep the light OFF after a power outage")
+            .withCategory("config"),
     colorPowerOnBehavior: () =>
-        e.enum("color_power_on_behavior", ea.STATE_SET, ["initial", "previous", "customized"]).withDescription("Power on behavior state"),
-    powerOnBehavior: () => e.enum("power_on_behavior", ea.ALL, ["off", "on", "previous"]).withDescription("Power on behavior state"),
+        e
+            .enum("color_power_on_behavior", ea.STATE_SET, ["initial", "previous", "customized"])
+            .withDescription("Power on behavior state")
+            .withCategory("config"),
+    powerOnBehavior: () =>
+        e.enum("power_on_behavior", ea.ALL, ["off", "on", "previous"]).withDescription("Power on behavior state").withCategory("config"),
     switchMode: () =>
-        e.enum("switch_mode", ea.STATE_SET, ["switch", "scene"]).withDescription("Sets the mode of the switch to act as a switch or as a scene"),
+        e
+            .enum("switch_mode", ea.STATE_SET, ["switch", "scene"])
+            .withDescription("Sets the mode of the switch to act as a switch or as a scene")
+            .withCategory("config"),
     switchMode2: () =>
         e
             .enum("switch_mode", ea.STATE_SET, ["switch", "curtain"])
-            .withDescription("Sets the mode of the switch to act as a switch or as a curtain controller"),
+            .withDescription("Sets the mode of the switch to act as a switch or as a curtain controller")
+            .withCategory("config"),
     lightMode: () =>
         e.enum("light_mode", ea.STATE_SET, ["normal", "on", "off", "flash"]).withDescription(`'Sets the indicator mode of l1.
         Normal: Orange while off and white while on.
@@ -537,6 +675,11 @@ export const valueConverter = {
         "button-switch": new Enum(2),
         "button2-switch": new Enum(3),
     }),
+    switchTypeButton: valueConverterBasic.lookup({
+        release: new Enum(0),
+        press: new Enum(1),
+    }),
+
     switchType2: valueConverterBasic.lookup({toggle: new Enum(0), state: new Enum(1), momentary: new Enum(2)}),
     backlightModeOffNormalInverted: valueConverterBasic.lookup({off: new Enum(0), normal: new Enum(1), inverted: new Enum(2)}),
     backlightModeOffLowMediumHigh: valueConverterBasic.lookup({off: new Enum(0), low: new Enum(1), medium: new Enum(2), high: new Enum(3)}),
@@ -1726,6 +1869,17 @@ const tuyaTz = {
             await entity.read("manuSpecificTuya3", ["switchType"]);
         },
     } satisfies Tz.Converter,
+    switch_type_button: {
+        key: ["switch_type_button"],
+        convertSet: async (entity, key, value, meta) => {
+            const switchType = utils.getFromLookup(value, {release: 0, press: 1});
+            await entity.write("manuSpecificTuya3", {switchType}, {disableDefaultResponse: true});
+            return {state: {[key]: value}};
+        },
+        convertGet: async (entity, key, meta) => {
+            await entity.read("manuSpecificTuya3", ["switchType"]);
+        },
+    } satisfies Tz.Converter,
     backlight_indicator_mode_1: {
         key: ["backlight_mode", "indicator_mode"],
         convertSet: async (entity, key, value, meta) => {
@@ -2007,6 +2161,17 @@ const tuyaFz = {
             }
         },
     } satisfies Fz.Converter<"manuSpecificTuya3", undefined, ["attributeReport", "readResponse"]>,
+    switch_type_button: {
+        cluster: "manuSpecificTuya3",
+        type: ["attributeReport", "readResponse"],
+        convert: (model, msg, publish, options, meta) => {
+            if (msg.data.switchType !== undefined) {
+                const lookup: KeyValue = {0: "release", 1: "press"};
+                utils.assertNumber(msg.data.switchType);
+                return {switch_type_button: lookup[msg.data.switchType]};
+            }
+        },
+    } satisfies Fz.Converter<"manuSpecificTuya3", undefined, ["attributeReport", "readResponse"]>,
     backlight_mode_low_medium_high: {
         cluster: "genOnOff",
         type: ["attributeReport", "readResponse"],
@@ -2145,6 +2310,57 @@ const tuyaFz = {
             }
         },
     } satisfies Fz.Converter<"manuSpecificTuya4", Tuya4, ["attributeReport", "readResponse"]>,
+    // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
+    TS011F_electrical_measurement: {
+        ...fz.electrical_measurement,
+        convert: (model, msg, publish, options, meta) => {
+            const result = (fz.electrical_measurement.convert(model, msg, publish, options, meta) as KeyValueAny) ?? {};
+            const lookup: KeyValueString = {
+                power: "activePower",
+                current: "rmsCurrent",
+                voltage: "rmsVoltage",
+            };
+
+            // Wait 5 seconds before reporting a 0 value as this could be an invalid measurement.
+            // https://github.com/Koenkk/zigbee2mqtt/issues/16709#issuecomment-1509599046
+            if (result) {
+                for (const key of ["power", "current", "voltage"]) {
+                    if (key in result) {
+                        const value = result[key];
+                        clearTimeout(globalStore.getValue(msg.endpoint, key));
+                        if (value === 0) {
+                            const configuredReporting = msg.endpoint.configuredReportings.find(
+                                (c) => c.cluster.name === "haElectricalMeasurement" && c.attribute.name === lookup[key],
+                            );
+                            const time = (configuredReporting ? configuredReporting.minimumReportInterval : 5) * 2 + 1;
+                            globalStore.putValue(
+                                msg.endpoint,
+                                key,
+                                setTimeout(() => {
+                                    const payload = {[key]: value};
+                                    // Device takes a lot of time to report power 0 in some cases. When current == 0 we can assume power == 0
+                                    // https://github.com/Koenkk/zigbee2mqtt/discussions/19680#discussioncomment-7868445
+                                    if (key === "current") {
+                                        payload.power = 0;
+                                    }
+                                    publish(payload);
+                                }, time * 1000),
+                            );
+                            delete result[key];
+                        }
+                    }
+                }
+            }
+
+            // Device takes a lot of time to report power 0 in some cases. When the state is OFF we can assume power == 0
+            // https://github.com/Koenkk/zigbee2mqtt/discussions/19680#discussioncomment-7868445
+            if (meta.state.state === "OFF") {
+                result.power = 0;
+            }
+
+            return result;
+        },
+    } satisfies Fz.Converter<"haElectricalMeasurement", undefined, ["attributeReport", "readResponse"]>,
 };
 export {tuyaFz as fz};
 
@@ -2816,6 +3032,7 @@ const tuyaModernExtend = {
             powerOnBehavior2?: boolean | ((manufacturerName: string) => boolean);
             switchType?: boolean | ((manufacturerName: string) => boolean);
             switchTypeCurtain?: boolean;
+            switchTypeButton?: boolean;
             backlightModeLowMediumHigh?: boolean;
             indicatorMode?: boolean | ((manufacturerName: string) => boolean);
             indicatorModeNoneRelayPos?: boolean;
@@ -2896,6 +3113,11 @@ const tuyaModernExtend = {
             fromZigbee.push(tuyaFz.switch_type_curtain);
             toZigbee.push(tuyaTz.switch_type_curtain);
             exposes.push(tuyaExposes.switchTypeCurtain());
+        }
+        if (args.switchTypeButton) {
+            fromZigbee.push(tuyaFz.switch_type_button);
+            toZigbee.push(tuyaTz.switch_type_button);
+            exposes.push(tuyaExposes.switchTypeButton());
         }
         if (backlightModeOffOn) {
             fromZigbee.push(tuyaFz.backlight_mode_off_on);
@@ -3054,9 +3276,19 @@ const tuyaModernExtend = {
         });
     },
     tuyaWeatherForecast(
-        args: {includeCurrentWeather?: boolean; numberOfForecastDays?: number; correctForNegativeValues?: boolean} = {},
+        args: {
+            includeCurrentWeather?: boolean;
+            numberOfForecastDays?: number;
+            correctForNegativeValues?: boolean;
+            weatherConditionMap?: TuyaWeatherCondition;
+        } = {},
     ): ModernExtend {
-        const {includeCurrentWeather = true, numberOfForecastDays = 3, correctForNegativeValues = false} = args;
+        const {
+            includeCurrentWeather = true,
+            numberOfForecastDays = 3,
+            correctForNegativeValues = false,
+            weatherConditionMap = M8ProTuyaWeatherCondition,
+        } = args;
 
         const tz_fileds = includeCurrentWeather ? ["temperature_0", "humidity_0", "condition_0"] : [];
 
@@ -3092,7 +3324,7 @@ const tuyaModernExtend = {
                 );
                 weather_values[TuyaWeatherID.Humidity].push("humidity_0" in meta.state ? (meta.state["humidity_0"] as number) : 0);
                 weather_values[TuyaWeatherID.Condition].push(
-                    "condition_0" in meta.state ? M8ProTuyaWeatherCondition[meta.state["condition_0"] as keyof typeof M8ProTuyaWeatherCondition] : 0,
+                    "condition_0" in meta.state ? weatherConditionMap[meta.state["condition_0"] as keyof typeof weatherConditionMap] : 0,
                 );
             } else {
                 buffer.writeUInt8(0x0, bOffset++);
@@ -3104,9 +3336,7 @@ const tuyaModernExtend = {
                 );
                 weather_values[TuyaWeatherID.Humidity].push(`humidity_${i}` in meta.state ? (meta.state[`humidity${i}`] as number) : 0);
                 weather_values[TuyaWeatherID.Condition].push(
-                    `condition_${i}` in meta.state
-                        ? M8ProTuyaWeatherCondition[meta.state[`condition_${i}`] as keyof typeof M8ProTuyaWeatherCondition]
-                        : 0,
+                    `condition_${i}` in meta.state ? weatherConditionMap[meta.state[`condition_${i}`] as keyof typeof weatherConditionMap] : 0,
                 );
             }